Remove duplicates in closed library strategies
