+++ src/main/java/org/folio/circulation/domain/policy/library/EndOfCurrentHoursStrategy.java
@@ -1,6 +1,5 @@
 package org.folio.circulation.domain.policy.library;
 
-import org.folio.circulation.AdjustingOpeningDays;
 import org.folio.circulation.support.HttpResult;
 import org.folio.circulation.support.ValidationErrorFailure;
 import org.folio.circulation.support.http.server.ValidationError;
@@ -8,35 +7,18 @@ import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 
 import java.util.Collections;
-import java.util.Objects;
 
-import static org.folio.circulation.domain.policy.library.ClosedLibraryStrategyUtils.failureForAbsentTimetable;
-
-public class EndOfCurrentHoursStrategy implements ClosedLibraryStrategy {
+public class EndOfCurrentHoursStrategy extends ShortTermLoansBaseStrategy {
 
   private final DateTime currentTime;
-  private final DateTimeZone zone;
 
-  public EndOfCurrentHoursStrategy(
-    DateTime currentTime, DateTimeZone zone) {
+  public EndOfCurrentHoursStrategy(DateTime currentTime, DateTimeZone zone) {
+    super(zone);
     this.currentTime = currentTime;
-    this.zone = zone;
   }
 
   @Override
-  public HttpResult<DateTime> calculateDueDate(DateTime requestedDate, AdjustingOpeningDays openingDays) {
-    Objects.requireNonNull(openingDays);
-    LibraryTimetable libraryTimetable =
-      LibraryTimetableConverter.convertToLibraryTimetable(openingDays, zone);
-
-    LibraryInterval requestedInterval = libraryTimetable.findInterval(requestedDate);
-    if (requestedInterval == null) {
-      return HttpResult.failed(failureForAbsentTimetable());
-    }
-    if (requestedInterval.isOpen()) {
-      return HttpResult.succeeded(requestedDate);
-    }
-
+  protected HttpResult<DateTime> calculateIfClosed(LibraryTimetable libraryTimetable, LibraryInterval requestedInterval) {
     LibraryInterval currentTimeInterval = libraryTimetable.findInterval(currentTime);
     if (currentTimeInterval == null) {
       return HttpResult.failed(errorForAbsentCurrentInterval());
@@ -47,6 +29,7 @@ public class EndOfCurrentHoursStrategy implements ClosedLibraryStrategy {
     return HttpResult.succeeded(currentTimeInterval.getEndTime());
   }
 
+
   private ValidationErrorFailure errorForAbsentCurrentInterval() {
     String message = "Unable to find current service point hours";
     return new ValidationErrorFailure(new ValidationError(message, Collections.emptyMap()));
@@ -56,6 +39,4 @@ public class EndOfCurrentHoursStrategy implements ClosedLibraryStrategy {
     String message = "Current service point hours are closed";
     return new ValidationErrorFailure(new ValidationError(message, Collections.emptyMap()));
   }
-
-
 }
